+++ capstone/yieldpay/programs/yieldpay/src/state/config.rs
@@ -8,7 +8,7 @@ pub const YIELD_MINT_SEED: &[u8] = b"YIELD";
 pub struct Config {
     pub admin:Pubkey,
     pub max_stake: u64,
-    pub min_deposit: u8,
+    pub min_deposit: u64,
     pub yield_mint:Pubkey,
     pub total_users: u64,
     pub total_merchants: u64,
